+++ urcl2isa/isa.py
@@ -1,5 +1,5 @@
 from typing import TYPE_CHECKING
-from urclToISA.operand import Operand 
+from operand import Operand 
 
 class Block():
     def __init__(self, URCL_labels:list[Operand]=[], code:list[str]=[]):
