--- urclToISA/__init__.py
@@ -1,6 +1,0 @@
-import urclToISA.operand
-import urclToISA.isa
-import urclToISA.instruction
-import urclToISA.program
-import urclToISA.translator
-import urclToISA.UTRX
